pyproject.toml
[tool.poetry]
name = "instant_repo"                             # TODO: change name
version = "0.15.5"                                # TODO: initialize to "0.1.0"
description = "template for personal projects"    # TODO: change description
authors = ["niklasbaier <niklas.baier@yahoo.de>"]
readme = "README.md"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "*"
mypy = "*"
types-requests = "*"
pre-commit = "*"
commitizen = "*"
ruff = "^0.4.4"
pytest-cov = "^5.0.0"

[tool.poetry.group.docs.dependencies]
mkdocs-material = "*"
termynal = "*"
mkdocs-git-authors-plugin = "*"
mkdocstrings = "*"
mkdocstrings-python = "*"
mkdocs-include-markdown-plugin = "*"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.15.5" # TODO: initialize to "0.1.0"
tag_format = "v$version"
version_files = [
    "pyproject.toml:^version",
    "src/instant_repo/__init__.py:^__version",
    "src/instant_repo/__init__.py:^__instant_repo_version", # TODO: remove
]
update_changelog_on_bump = true
changelog_incremental = true
changelog_merge_prerelease = true
bump_message = "bump: $current_version -> $new_version"
